Move stat tracker methods into statistic-specific modules
